+++ src/Module.php
@@ -195,6 +195,13 @@ final class Module extends \luya\admin\base\Module implements CoreModuleInterfac
     public $fileDefaultInlineDisposition = false;
 
     /**
+     * @var boolean Defines whether Api Users can access a method which is not protected from the permission system. When working with JWT or SPA
+     * applications this should be disabled.
+     * @since 2.2.0
+     */
+    public $apiUserAllowActionsWithoutPermissions = false;
+
+    /**
      * @var array A configuration array with all tags shipped by default with the admin module.
      */
     public $tags = [
@@ -224,7 +231,6 @@ final class Module extends \luya\admin\base\Module implements CoreModuleInterfac
         'api-admin-proxy' => 'luya\admin\apis\ProxyController',
         'api-admin-config' => 'luya\admin\apis\ConfigController',
         'api-admin-queuelog' => 'luya\admin\apis\QueueLogController',
-        'api-admin-userrequest' => 'luya\admin\apis\UserRequestController',
     ];
 
     /**
@@ -243,6 +249,9 @@ final class Module extends \luya\admin\base\Module implements CoreModuleInterfac
      */
     public $moduleMenus = [];
     
+    /**
+     * @inheritDoc
+     */
     public static function onLoad()
     {
         self::registerTranslation('admin*', '@admin/messages', [