+++ src/apis/StorageController.php
@@ -23,6 +23,7 @@ use luya\admin\models\TagRelation;
 use luya\admin\traits\TaggableTrait;
 use luya\admin\storage\BaseFileSystemStorage;
 use luya\admin\events\FileEvent;
+use yii\base\Action;
 
 /**
  * Filemanager and Storage API.
@@ -43,14 +44,9 @@ class StorageController extends RestController
      */
     const PERMISSION_ROUTE = 'admin/storage/index';
     
-    /**
-     * Flush the storage caching data.
-     */
-    protected function flushApiCache($folderId = 0, $page = 0)
+    public function permissionRoute(Action $action)
     {
-        Yii::$app->storage->flushArrays();
-        $this->deleteHasCache('storageApiDataFolders');
-        $this->deleteHasCache(['storageApiDataFiles', (int) $folderId, (int) $page]);
+        return self::PERMISSION_ROUTE;
     }
     
     // DATA READERS
@@ -99,8 +95,6 @@ class StorageController extends RestController
             'query' => $query,
         ]);
     }
-    
-    // ACTIONS
     
     /**
      * Toggle Tags for a given file.
@@ -274,8 +268,6 @@ class StorageController extends RestController
      */
     public function actionFilemanagerUpdateCaption()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         $fileId = Yii::$app->request->post('id', false);
         $captionsText = Yii::$app->request->post('captionsText', false);
         $pageId = Yii::$app->request->post('pageId', 0);
@@ -306,7 +298,6 @@ class StorageController extends RestController
      */
     public function actionImageFilter()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
         $image = Yii::$app->storage->createImage(Yii::$app->request->post('fileId', null), Yii::$app->request->post('filterId', null));
         if ($image) {
             return [
@@ -339,8 +330,6 @@ class StorageController extends RestController
      */
     public function actionFileReplace()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         $fileId = Yii::$app->request->post('fileId', false);
         $pageId = Yii::$app->request->post('pageId', 0);
         Yii::warning('replace request for file id' . $fileId, __METHOD__);
@@ -433,8 +422,6 @@ class StorageController extends RestController
     */
     public function actionFilesUpload()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         foreach ($_FILES as $k => $file) {
             if ($file['error'] !== UPLOAD_ERR_OK) {
                 Yii::$app->response->setStatusCode(422, 'Data Validation Failed.');
@@ -466,8 +453,6 @@ class StorageController extends RestController
      */
     public function actionFilemanagerMoveFiles()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         $toFolderId = Yii::$app->request->post('toFolderId', 0);
         $fileIds = Yii::$app->request->post('fileIds', []);
         
@@ -489,7 +474,6 @@ class StorageController extends RestController
      */
     public function actionFilemanagerRemoveFiles()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
         $pageId = Yii::$app->request->post('pageId', 0);
         $folderId = Yii::$app->request->post('folderId', 0);
         foreach (Yii::$app->request->post('ids', []) as $id) {
@@ -530,8 +514,6 @@ class StorageController extends RestController
      */
     public function actionFolderDelete($folderId)
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         // find all subfolders
         $matchingChildFolders = StorageFolder::find()->where(['parent_id' => $folderId])->asArray()->all();
         foreach ($matchingChildFolders as $matchingChildFolder) {
@@ -564,8 +546,6 @@ class StorageController extends RestController
      */
     public function actionFolderUpdate($folderId)
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         $model = StorageFolder::findOne($folderId);
         if (!$model) {
             return false;
@@ -584,13 +564,21 @@ class StorageController extends RestController
      */
     public function actionFolderCreate()
     {
-        $this->checkRouteAccess(self::PERMISSION_ROUTE);
-        
         $folderName = Yii::$app->request->post('folderName', null);
         $parentFolderId = Yii::$app->request->post('parentFolderId', 0);
         $response = Yii::$app->storage->addFolder($folderName, $parentFolderId);
         $this->flushApiCache();
         
         return $response;
+    }
+    
+    /**
+     * Flush the storage caching data.
+     */
+    protected function flushApiCache($folderId = 0, $page = 0)
+    {
+        Yii::$app->storage->flushArrays();
+        $this->deleteHasCache('storageApiDataFolders');
+        $this->deleteHasCache(['storageApiDataFiles', (int) $folderId, (int) $page]);
     }
 }