+++ tests/data/controllers/TestController.php
@@ -2,10 +2,17 @@
 
 namespace luya\admin\tests\data\controllers;
 
-use luya\rest\Controller;
+use luya\admin\base\RestController;
 
-class TestController extends Controller
+class TestController extends RestController
 {
+    public $authOptional = ['no-auth'];
+
+    public function actionNoAuth()
+    {
+        return true;
+    }
+
     public function actionBarFoo()
     {
         return true;